requirements.txt
langchain
pypdf
tiktoken
pinecone-client
cohere
python-dotenv
promptwatch
chainlit
flask
Jinja2